refactor: centralize PlayerTypeView
